# api/core.openapi.yaml
basePath: /
definitions:
...
  InactiveUserListResponse:
    properties:
      days:
        type: integer
      items:
        items:
          $ref: '#/definitions/UserActivityRow'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_login:
        type: string
      last_login_local_date:
        type: string
      last_name:
        type: string
      login_count:
        type: integer
      login_day_count:
        type: integer
      login_streak:
        type: integer
      middle_name:
        type: string
      organization_id:
        type: string
      status:
        $ref: '#/definitions/UserStatus'
      timezone:
        type: string
      type:
        $ref: '#/definitions/UserType'
    type: object
  UserActivityAnalytics:
    properties:
      active_users:
        type: integer
      average_logins_per_active_day:
        type: number
        x-nullable: true
      generated_at:
        type: string
      lifetime_active_days:
        type: integer
      lifetime_login_count:
        type: integer
      monthly_active_days:
        type: integer
      monthly_active_users:
        type: integer
      monthly_login_count:
        type: integer
      total_users:
        type: integer
      trend:
        items:
          $ref: '#/definitions/UserActivityTrendPoint'
        type: array
      users_ever_logged_in:
        type: integer
      window_active_days:
        type: integer
      window_days:
        type: integer
      window_login_count:
        type: integer
    type: object
  UserActivityListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/UserActivityRow'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  UserActivityRow:
    properties:
      average_logins_per_active_day:
        type: number
        x-nullable: true
      consecutive_logins:
        description: Deprecated Phase 1 compatibility alias.
        type: integer
      current_streak:
        type: integer
      effective_timezone:
        type: string
      email:
        type: string
      last_login:
        type: string
        x-nullable: true
      last_login_local_date:
        type: string
        x-nullable: true
      inactive_days:
        description: User-local calendar days since the last login; null for never-logged-in users.
        type: integer
        x-nullable: true
      login_count:
        type: integer
      login_day_count:
        type: integer
      name:
        type: string
      user_id:
        type: string
    type: object
  UserActivityTrendPoint:
    properties:
      active_users:
        type: integer
      date:
        type: string
      login_count:
        type: integer
    type: object
...
  UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/User'
        type: array
      total:
        type: integer
    type: object
...
  UserStatus:
    enum:
    - PENDING
    - ACTIVE
    - INACTIVE
    - SUSPENDED
    - MERGED
    type: string
    x-enum-varnames:
    - UserPending
    - UserActive
    - UserInactive
    - UserSuspended
    - UserMerged
  UserType:
    enum:
    - CLUSTER_LOCAL
    - VIRTUAL
    - SYSTEM
    type: string
    x-enum-varnames:
    - UserTypeClusterLocal
    - UserTypeVirtual
    - UserTypeSystem
...
  mergeUsersRequest:
    properties:
      retiring_user_id:
        type: string
      surviving_user_id:
        type: string
    type: object
...
  userNameUpdateRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'REST API for Apache Custos: organizations, users, projects, compute
    clusters, allocations, audit traces, and the privilege/role authorization layer.
    OIDC bearer auth required. Connector endpoints live under /connectors/<name>/.'
  title: Apache Custos Core API
  version: 0.2.0
paths:
...
  /users:
    get:
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserListResponse'
      security:
      - BearerAuth: []
      summary: List users (paginated)
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - Users
  /users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Name fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userNameUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user's name
      tags:
      - Users
...
  /users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status patch
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              $ref: '#/definitions/UserStatus'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user's status
      tags:
      - Users
...
  /users/activity:
    get:
      description: Returns complete login activity summaries with server-side search,
        ordering, and pagination.
      parameters:
      - description: Case-insensitive name or email search
        in: query
        name: query
        type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      - description: 'Sort field: name, last_login, login_count, login_day_count,
          or current_streak'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: asc or desc'
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserActivityListResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List user activity (paginated)
      tags:
      - Users
  /users/activity/analytics:
    get:
      description: Returns lifetime, current-month, and rolling-window login metrics
        plus a daily trend. Rolling and monthly boundaries use each daily fact's recorded
        user-local IANA timezone.
      parameters:
      - description: 'Rolling window in days: 7, 30, or 90 (default 30)'
        in: query
        name: window
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserActivityAnalytics'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user activity analytics
      tags:
      - Users
  /users/{id}/activity/analytics:
    get:
      description: Returns login engagement metrics restricted to the explicitly selected user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Rolling window in days: 7, 30, or 90 (default 30)'
        in: query
        name: window
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserActivityAnalytics'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get analytics for one selected user
      tags:
      - Users
  /users/inactive:
    get:
      description: Returns users inactive for at least the requested number of user-local
        calendar days. Never-logged-in users are included first.
      parameters:
      - description: Minimum inactive calendar days (default 7, max 3650)
        in: query
        name: days
        type: integer
      - description: Case-insensitive name or email search
        in: query
        name: query
        type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/InactiveUserListResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List inactive users (paginated)
      tags:
      - Users
  /users/merge:
    post:
      consumes:
      - application/json
      description: Merges the retiring user into the surviving user; the surviving
        record is returned.
      parameters:
      - description: Merge payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mergeUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Merge two users
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"